api/Beek/requirements.txt
asgiref==3.8.1
boto3==1.35.24
Django==4.2.16
django-filter==24.3
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
setuptools==72.1.0
sqlparse==0.5.1
stripe==10.8.0
uritemplate==4.1.1
daphne==4.1.2
channels==4.1.0
channels-redis==4.2.0
celery==5.4.0
django-storages==1.14.4
SQLAlchemy==2.0.32
openpyxl==3.1.5
PyMuPDF==1.25.1